# src/main/java/com/examen/hei/model/CashAccount.java
package com.examen.hei.model;

public class CashAccount extends FinancialAccount {
    private String type;

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }
}